Refactored the design for liquidity page
